server/pyproject.toml
[tool.poetry]
name = "event-capture-server"
version = "0.1.0"
description = "FastAPI service for event capture with MongoDB storage"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pymongo = "^4.8.0"
pydantic = "^2.8.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
